refactor: 레이아웃 컴포넌트를 organism에서 component하위로 이동
