+++ src/components/sections/InterviewSection.tsx
@@ -1,6 +1,6 @@
+import SlideUpSection from "@/components/layouts/SlideUpSection";
+import Content from "@/components/layouts/Content";
 import InterviewItem from "@/components/molecules/InterviewItem";
-import SlideUpSection from "@/components/organisms/layouts/SlideUpSection";
-import Content from "@/components/organisms/layouts/Content";
 
 const InterviewSection = () => {
   return (
