+++ src/components/sections/IntroductionSection.tsx
@@ -1,9 +1,9 @@
 import SlideAnimation from "@/components/animations/SlideAnimation";
+import Section from "@/components/layouts/Section";
 import Display1 from "@/components/atoms/typography/Display1";
 import Title1 from "@/components/atoms/typography/Title1";
 import Heading1 from "@/components/atoms/typography/Heading1";
 import ContactGroup from "@/components/molecules/ContactGroup";
-import Section from "@/components/organisms/layouts/Section";
 
 const IntroductionSection = () => {
   return (
